refactor: add BoxStackSeries component
