+++ src/component/boxSeries.ts
@@ -1,50 +1,26 @@
 import Component from './component';
 import { RectModel, BoxSeriesModel, ClipRectAreaModel } from '@t/components/series';
-import {
-  ChartState,
-  ChartType,
-  StackData,
-  SeriesData,
-  StackGroupData,
-  StackDataType,
-  AxisData
-} from '@t/store/store';
+import { ChartState, ChartType, SeriesData } from '@t/store/store';
 import {
   BoxSeriesType,
   BoxSeriesDataType,
   RangeDataType,
   BarChartOptions,
-  ColumnChartOptions,
-  StackType,
-  StackInfo
+  ColumnChartOptions
 } from '@t/options';
 import { first, last, includes } from '@src/helpers/utils';
 import { TooltipData } from '@t/components/tooltip';
 
 type DrawModels = BoxSeriesModel | ClipRectAreaModel | RectModel;
 
-type SizeKey = 'width' | 'height';
+export type SizeKey = 'width' | 'height';
 
-type SeriesRawData = BoxSeriesType<BoxSeriesDataType>[];
-
-interface StackSeriesModelParamType {
-  stackData: StackData;
-  colors: string[];
-  valueLabels: string[];
-  tickDistance: number;
-  offsetSizeKey: SizeKey;
-  stackGroup?: { count: number; index: number };
-}
+export type SeriesRawData = BoxSeriesType<BoxSeriesDataType>[];
 
 export enum BoxType {
   BAR = 'bar',
   COLUMN = 'column'
 }
-
-export const STACK_TYPES = {
-  NORMAL: 'normal',
-  PERCENT: 'percent'
-};
 
 const PADDING = {
   TB: 15, // top & bottom
@@ -53,10 +29,6 @@ const PADDING = {
 
 function isRangeData(value): value is RangeDataType {
   return Array.isArray(value);
-}
-
-function isGroupStack(rawData: StackDataType): rawData is StackGroupData {
-  return !Array.isArray(rawData);
 }
 
 export function isBoxSeries(seriesName: ChartType): seriesName is BoxType {
@@ -75,8 +47,6 @@ export default class BoxSeries extends Component {
   isBar = true;
 
   name = BoxType.BAR;
-
-  stack!: StackInfo;
 
   initialize({ name }: { name: BoxType }) {
     this.type = 'series';
@@ -100,12 +70,21 @@ export default class BoxSeries extends Component {
     const { layout, series, theme, axes, categories } = chartState;
 
     this.rect = layout.plot;
-    this.stack = series[this.name]!.stack!;
-
     const { colors } = theme.series;
     const seriesData = series[this.name]!;
 
-    const seriesModels: BoxSeriesModel[] = this.renderSeriesModel(seriesData, colors, axes);
+    const valueAxis = this.isBar ? 'xAxis' : 'yAxis';
+    const labelAxis = this.isBar ? 'yAxis' : 'xAxis';
+    const anchorSizeKey = this.isBar ? 'height' : 'width';
+    const offsetSizeKey = this.isBar ? 'width' : 'height';
+    const tickDistance = this.rect[anchorSizeKey] / axes[labelAxis].validTickCount;
+    const seriesModels: BoxSeriesModel[] = this.renderSeriesModel(
+      seriesData,
+      colors,
+      axes[valueAxis].labels,
+      tickDistance,
+      offsetSizeKey
+    );
 
     const tooltipData: TooltipData[] = this.makeTooltipData(seriesData, colors, categories);
 
@@ -119,43 +98,14 @@ export default class BoxSeries extends Component {
     }));
   }
 
-  private renderSeriesModel(
+  renderSeriesModel(
     seriesData: SeriesData<BoxType>,
-    colors: string[],
-    axes: Record<string, AxisData>
-  ) {
-    const valueAxis = this.isBar ? 'xAxis' : 'yAxis';
-    const labelAxis = this.isBar ? 'yAxis' : 'xAxis';
-    const anchorSizeKey = this.isBar ? 'height' : 'width';
-    const offsetSizeKey = this.isBar ? 'width' : 'height';
-    const tickDistance = this.rect[anchorSizeKey] / axes[labelAxis].validTickCount;
-
-    if (this.stack) {
-      return this.renderStackSeriesModel(
-        seriesData,
-        colors,
-        axes[valueAxis].labels,
-        tickDistance,
-        offsetSizeKey
-      );
-    }
-
-    return this.renderBoxSeriesModel(
-      seriesData.data,
-      colors,
-      axes[valueAxis].labels,
-      tickDistance,
-      offsetSizeKey
-    );
-  }
-
-  renderBoxSeriesModel(
-    seriesRawData: SeriesRawData,
     colors: string[],
     valueLabels: string[],
     tickDistance: number,
     offsetSizeKey: SizeKey
   ): BoxSeriesModel[] {
+    const seriesRawData = seriesData.data;
     const minValue = Number(first(valueLabels));
     const offsetAxisLength = this.rect[offsetSizeKey];
     const axisValueRatio = offsetAxisLength / (Number(last(valueLabels)) - minValue);
@@ -190,27 +140,7 @@ export default class BoxSeries extends Component {
     });
   }
 
-  private renderStackSeriesModel(
-    seriesData: SeriesData<BoxType>,
-    colors: string[],
-    valueLabels: string[],
-    tickDistance: number,
-    offsetSizeKey: SizeKey
-  ): BoxSeriesModel[] {
-    const stackData = seriesData.stackData!;
-
-    return isGroupStack(stackData)
-      ? this.makeStackGroupSeriesModel(
-          seriesData,
-          [...colors],
-          valueLabels,
-          tickDistance,
-          offsetSizeKey
-        )
-      : this.makeStackSeriesModel({ stackData, colors, valueLabels, tickDistance, offsetSizeKey });
-  }
-
-  private renderClipRectAreaModel(): ClipRectAreaModel {
+  protected renderClipRectAreaModel(): ClipRectAreaModel {
     return {
       type: 'clipRectArea',
       x: 0,
@@ -220,7 +150,7 @@ export default class BoxSeries extends Component {
     };
   }
 
-  private renderRect(seriesModel): RectModel[] {
+  protected renderRect(seriesModel): RectModel[] {
     return seriesModel.map(data => {
       const { x, y, width, height, color } = data;
 
@@ -258,10 +188,6 @@ export default class BoxSeries extends Component {
   ): TooltipData[] {
     const seriesRawData = seriesData.data;
 
-    if (this.stack) {
-      return this.getStackTooltip(seriesData, colors, categories);
-    }
-
     return seriesRawData.flatMap(({ name, data }, index) =>
       data.map((value, dataIdx) => ({
         label: name,
@@ -272,54 +198,11 @@ export default class BoxSeries extends Component {
     );
   }
 
-  private getStackTooltip(
-    seriesRaw: SeriesData<BoxType>,
-    colors: string[],
-    categories?: string[]
-  ): TooltipData[] {
-    const seriesRawData = seriesRaw.data;
-    const stackData = seriesRaw.stackData!;
-
-    return isGroupStack(stackData)
-      ? this.makeGroupStackTooltipData(seriesRawData, stackData, colors, categories)
-      : this.makeStackTooltipData(seriesRawData, stackData, colors, categories);
-  }
-
-  private makeGroupStackTooltipData(
-    seriesRawData: SeriesRawData,
-    stackData: StackGroupData,
-    colors: string[],
-    categories?: string[]
-  ) {
-    return Object.keys(stackData).flatMap((groupId, groupIdx) => {
-      const filtered = seriesRawData.filter(({ stackGroup }) => stackGroup === groupId);
-      const groupColors = colors.splice(groupIdx, filtered.length);
-
-      return this.makeStackTooltipData(seriesRawData, stackData[groupId], groupColors, categories);
-    });
-  }
-
-  private makeStackTooltipData(
-    seriesRawData: SeriesRawData,
-    stackData: StackData,
-    colors: string[],
-    categories?: string[]
-  ) {
-    return stackData.flatMap(({ values }, index) =>
-      values.map((value, seriesIndex) => ({
-        label: seriesRawData[seriesIndex].name,
-        color: colors[seriesIndex],
-        value,
-        category: categories?.[index]
-      }))
-    );
-  }
-
   private getTooltipValue(value) {
     return isRangeData(value) ? `${value[0]} ~ ${value[1]}` : value;
   }
 
-  private getTotalOfPrevValues(values, currentIndex, included = false) {
+  protected getTotalOfPrevValues(values, currentIndex, included = false) {
     return values.reduce((a, b, idx) => {
       const isPrev = included ? idx <= currentIndex : idx < currentIndex;
 
@@ -329,88 +212,5 @@ export default class BoxSeries extends Component {
 
       return a;
     }, 0);
-  }
-
-  makeStackSeriesModel({
-    stackData,
-    colors,
-    valueLabels,
-    tickDistance,
-    offsetSizeKey,
-    stackGroup = {
-      count: 1,
-      index: 0
-    }
-  }: StackSeriesModelParamType) {
-    const seriesModels: BoxSeriesModel[] = [];
-    const offsetAxisLength = this.rect[offsetSizeKey];
-    const columnWidth = (tickDistance - this.padding * 2) / stackGroup.count;
-    const stackType: StackType = this.stack.type;
-
-    stackData.forEach(({ values, sum }, index) => {
-      const seriesPos = index * tickDistance + this.padding + columnWidth * stackGroup.index;
-
-      values.forEach((value, seriesIndex) => {
-        const color = colors[seriesIndex];
-        const beforeValueSum = this.getTotalOfPrevValues(values, seriesIndex, !this.isBar);
-        let barLength, startPosition;
-
-        if (stackType === STACK_TYPES.PERCENT) {
-          barLength = (value / sum) * offsetAxisLength;
-          startPosition = (beforeValueSum / sum) * offsetAxisLength;
-        } else {
-          const offsetValue = Number(last(valueLabels)) - Number(first(valueLabels));
-          const axisValueRatio = offsetAxisLength / offsetValue;
-
-          barLength = value * axisValueRatio;
-          startPosition = beforeValueSum * axisValueRatio;
-        }
-
-        seriesModels.push({
-          type: 'box',
-          color,
-          width: this.isBar ? barLength : columnWidth,
-          height: this.isBar ? columnWidth : barLength,
-          x: this.isBar ? startPosition : seriesPos,
-          y: this.isBar ? seriesPos : offsetAxisLength - startPosition
-        });
-      });
-    });
-
-    return seriesModels;
-  }
-
-  makeStackGroupSeriesModel(
-    seriesRaw: SeriesData<BoxType>,
-    colors: string[],
-    valueLabels: string[],
-    tickDistance: number,
-    offsetSizeKey: SizeKey
-  ) {
-    const stackGroupData = seriesRaw.stackData as StackGroupData;
-    const seriesRawData = seriesRaw.data;
-    const stackGroupIds = Object.keys(stackGroupData);
-    let seriesModels: BoxSeriesModel[] = [];
-
-    stackGroupIds.forEach((groupId, index) => {
-      const filtered = seriesRawData.filter(({ stackGroup }) => stackGroup === groupId);
-
-      seriesModels = [
-        ...seriesModels,
-        ...this.makeStackSeriesModel({
-          stackData: stackGroupData[groupId],
-          colors: colors.splice(index, filtered.length),
-          valueLabels,
-          tickDistance,
-          offsetSizeKey,
-          stackGroup: {
-            count: stackGroupIds.length,
-            index
-          }
-        })
-      ];
-    });
-
-    return seriesModels;
   }
 }
